feat(): install and create base for routing
resolve #11
